seed scripts in their own file. script will be added to build command in package json
