+++ backend/data/db.js
@@ -1,22 +1,21 @@
 const { Sequelize } = require('sequelize');
-const Property = require('../models/Property');
 
 const sequelize = new Sequelize(
-  'Header',
-  process.env.MYSQL_USERNAME,
-  process.env.MYSQL_PASSWORD,
-  {
-    host: 'localhost',
-    dialect: 'mysql',
-  }
+	'Header',
+	process.env.MYSQL_USERNAME,
+	process.env.MYSQL_PASSWORD,
+	{
+		host: 'localhost',
+		dialect: 'mysql',
+	}
 );
 
 (async () => {
-  try {
-    await sequelize.authenticate();
-  } catch (error) {
-    console.log('ERROR IN db.js', error);
-  }
+	try {
+		await sequelize.authenticate();
+	} catch (error) {
+		console.log('ERROR IN db.js', error);
+	}
 })();
 
 module.exports = sequelize;
